requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Audio processing
pydub>=0.25.1
ffmpeg-python>=0.2.0

# Speaker diarization
pyannote.audio>=3.1.1
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.26.0

# Fast downloads
hf_transfer>=0.1.3
requests>=2.31.0 